+++ reworkedcamera/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+### 1.3.0
+
+- Changed the range of distances
+- Changed the collision logic
+- Added client auto-detection
+
 ### 1.1.0
 
 - Initial release